Pathways/requirements.txt
dash
dash-core-components
dash-html-components
dash-renderer
dash-table
Flask-SQLAlchemy
python-dotenv
sqlalchemy-citext

